inspector: add missing initialization
Add missing initialization reported by coverity. With
proper usage it looks like it will be initialized but
still good to have it initialized to a known state in
case that changes or invalide usage.

The method used to create object ids newObjectId() starts
at 1 so initializing to 0 should be safe and what we get
now when compiled in a way that auto initializes (for example
debug)

Signed-off-by: Michael Dawson <mdawson@devrus.com>

PR-URL: https://github.com/nodejs/node/pull/43254
Reviewed-By: Darshan Sen <raisinten@gmail.com>
Reviewed-By: Minwoo Jung <nodecorelab@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>